+++ packages/docusaurus-plugin-openapi-docs/src/markdown/createStatusCodes.ts
@@ -54,6 +54,7 @@ export default function json2xml(o: any, tab: any) {
 }
 
 interface Props {
+  id?: string;
   label?: string;
   responses: ApiItem["responses"];
 }
@@ -255,7 +256,7 @@ export function createExampleFromSchema(schema: any, mimeType: string) {
   return undefined;
 }
 
-export function createStatusCodes({ label, responses }: Props) {
+export function createStatusCodes({ label, id, responses }: Props) {
   if (responses === undefined) {
     return undefined;
   }
@@ -271,6 +272,7 @@ export function createStatusCodes({ label, responses }: Props) {
         children: [
           create("ApiTabs", {
             label,
+            id,
             children: codes.map((code) => {
               const responseHeaders: any = responses[code].headers;
               return create("TabItem", {
