+++ packages/docusaurus-theme-openapi-docs/src/theme/OperationTabs/_OperationTabs.scss
@@ -21,8 +21,7 @@
   margin-top: 0 !important;
   margin-right: 0.5rem;
   font-weight: var(--ifm-font-weight-bold);
-  font-size: 10px;
-  text-transform: uppercase;
+  font-size: 12px;
   white-space: nowrap;
   transition: 300ms;
